Handle critical exception during password reset
Started by adding the workaround as described in dotnet/aspnetcore#23638
and faced the same unhandled exception. Updating all related libraries
did not solve the issue.

I.e. password reset was working but user was left rather
clueless as the flow ended with an unhandled exception.

Due to the lack of a global exception handling (dotnet/aspnetcore#13452)
this change also includeds a fake logger which reacts
on critical logs in WebAssemblyRenderer and reloads the main page.

Closes #35