feat: Floor mic values to closest reference step (#137)
E-Test has its own range of MIC-values but this technique is always
compared with and evaluated by the MIC ranges of the reference
methode broth microdilution. Therefore e.g. only 0.06 and 0.125 do
matter as endpoints of the readout and serve for a breakpoint
evaluation.

Everything in between these values do not exsist in the ref methode. that is why if these strains do not reach the next MIC of 0.125 they have to be evaluated as susceptible.

E.g. a value of 0.064 should be interpreted as 0.06. When a breakpoint
is set as >0.06, 0.06 and 0.064 should be reported as susceptible.

Refs: #131